@@ -1,6 +1,6 @@
 [metadata]
 name = seismic
-version = 0.3.12
+version = 0.5.3
 author = Peter Makus, Christoph Sens-Schönfelder, and the SeisMIC development team
 author_email = makus@gfz-potsdam.de
 description = Seismological Monitoring using Interferometric Concepts
@@ -16,24 +16,25 @@ classifiers =
     Operating System :: OS Independent
     Development Status :: 4 - Beta
     Topic :: Utilities
-keywords = Seismology, Ambient Noise, Earth Sciences, Lithosphere, Time-dependent dynamics, Inteferometry
+keywords = Seismology, Ambient Noise, Earth Sciences, Environmental Seismology, Volcano Seismology, Critical Zone, Time-dependent dynamics, Inteferometry, Monitoring, Structural Health
 
 
 [options]
 package_dir =
     = src
-python_requires = >=3.7
+python_requires = >=3.10
 install_requires =
-    geographiclib
-    h5py
-    matplotlib
-    mpi4py
-    numpy
-    obspy >=1.3.1
+    geographiclib==2.0
+    h5py <=3.9.0
+    matplotlib <=3.7.2
+    mpi4py <=3.1.4
+    numpy <=1.25.2
+    obspy<=1.4.0, >=1.3.1
+
     pip
     prov
     pyyaml
-    scipy
+    scipy <=1.11.2
     flake8
     tqdm
 tests_require =
@@ -48,6 +49,8 @@ docs =
     pydata-sphinx-theme
     sphinx-copybutton
     sphinxcontrib-mermaid
+    jinja2<3.1
+    jupyter
 tests =
     pytest
     py
