@@ -10,7 +10,7 @@ Module that contains functions for preprocessing in the time domain
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Tuesday, 20th July 2021 03:24:01 pm
-Last Modified: Friday, 14th July 2023 10:29:04 am
+Last Modified: Thursday, 29th June 2023 11:15:56 am
 '''
 from copy import deepcopy
 
@@ -45,26 +45,104 @@ def clip(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
     :rtype: numpy.ndarray
     :return: clipped time series data
     """
-    stds = np.nanstd(A, axis=0)
-    for ind in range(A.shape[1]):
+    stds = np.nanstd(A, axis=1)
+    for ind in range(A.shape[0]):
         ts = args['std_factor']*stds[ind]
-        A[A[:, ind] > ts, ind] = ts
-        A[A[:, ind] < -ts, ind] = -ts
+        A[ind, A[ind, :] > ts] = ts
+        A[ind, A[ind, :] < -ts] = -ts
     return A
 
 
 def detrend(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
     """
-    Remove trend from data
+    Detrend wrapper. Can call scipy.signal.detrend or a faster custom function.
+
+    :param A: 1 or 2D array
+    :type A: np.ndarray
+    :param args: arguments for detrend function
+    :type args: dict
+    :param params: params file
+    :type params: dict
+    :raises ValueError: For unknown detrend method
+    :raises ValueError: For unknown detrend type
+    :return: detrended data
+    :rtype: np.ndarray
     """
-    try:
-        A[np.logical_not(np.isnan(A))] = sp_detrend(
-            A[np.logical_not(np.isnan(A))], axis=0, overwrite_data=True,
-            **args)
-    except (ZeroDivisionError, ValueError):
-        # When the array is nothing but nans
+    method = args.get('method', 'qr')
+    if method == 'scipy':
+        return detrend_scipy(A, args, params)
+    elif method == 'qr':
+        detrend_type = args.get('type', 'linear')
+        if detrend_type == 'linear':
+            return detrendqr(A)
+        elif detrend_type == 'constant':
+            return demean(A)
+        else:
+            raise ValueError('Unknown detrend type')
+    else:
+        raise ValueError('Unknown detrend method')
+
+
+def detrend_scipy(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
+    """
+    Remove trend from data
+
+    .. seealso:: :func:`scipy.signal.detrend` for input arguments
+    """
+    if np.all(np.isnan(A)):
         return A
+    for ii in range(A.shape[0]):
+        try:
+            A[ii, ~np.isnan(A[ii])] = sp_detrend(
+                A[ii, ~np.isnan(A[ii])], axis=-1, overwrite_data=True,
+                **args)
+        except ZeroDivisionError:
+            # When the line is nothing but nans
+            pass
     return A
+
+
+def detrendqr(data: np.ndarray) -> np.ndarray:
+    """
+    Remove trend from data using QR decomposition. Faster than
+    scipy.signal.detrend. Shamelessly adapted from
+    NoisePy
+
+    .. seealso:: Adapted from `NoisePy <https://doi.org/10.1785/0220190364>`_
+
+    :param data: 1 or 2D array
+    :type data: np.ndarray
+    :raises ValueError: For data with more than 2 dimensions
+    :return: detrended data
+    :rtype: np.ndarray
+    """
+    npts = data.shape[-1]
+    X = np.ones((npts, 2))
+    X[:, 0] = np.arange(0, npts, dtype=np.float32) / npts
+    Q, R = np.linalg.qr(X)
+    rq = np.dot(np.linalg.inv(R), Q.transpose())
+    if data.ndim == 1:
+        coeff = np.dot(rq, data)
+        data = data - np.dot(X, coeff)
+    elif data.ndim == 2:
+        for ii in range(data.shape[0]):
+            coeff = np.dot(rq, data[ii])
+            data[ii] = data[ii] - np.dot(X, coeff)
+    else:
+        raise ValueError('data must be 1D or 2D')
+    return data
+
+
+def demean(data: np.ndarray) -> np.ndarray:
+    """
+    Demean data
+
+    :param data: 1 or 2D array
+    :type data: np.ndarray
+    :return: demeaned data
+    :rtype: np.ndarray
+    """
+    return data - np.nanmean(data, axis=-1, keepdims=True)
 
 
 def mute(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
@@ -125,34 +203,34 @@ def mute(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
 
     # return zeros if length of traces is shorter than taper
     ntap = int(args['taper_len']*params['sampling_rate'])
-    if A.shape[0] <= ntap:
+    if A.shape[1] <= ntap:
         return np.zeros_like(A)
 
     # filter if asked to
     if 'filter' in list(args.keys()):
         C = TDfilter(A, args['filter'], params)
     else:
-        C = deepcopy(A)
+        C = A
 
     # calculate envelope
     D = np.abs(C)
 
     # calculate threshold
     if 'threshold' in list(args.keys()):
-        thres = np.zeros(A.shape[1]) + args['threshold']
+        thres = np.zeros(A.shape[0]) + args['threshold']
     elif 'std_factor' in list(args.keys()):
-        thres = np.std(C, axis=0) * args['std_factor']
+        thres = np.std(C, axis=1) * args['std_factor']
     else:
-        thres = np.std(C, axis=0)
+        thres = np.std(C, axis=1)
 
     # calculate mask
     mask = np.ones_like(D)
-    mask[D > np.tile(np.atleast_2d(thres), (A.shape[0], 1))] = 0
+    mask[D > np.tile(thres, (D.shape[1], 1)).T] = 0
     # extend the muted segments to make sure the whole segment is zero after
     if args['extend_gaps']:
         tap = np.ones(ntap)/ntap
-        for ind in range(A.shape[1]):
-            mask[:, ind] = np.convolve(mask[:, ind], tap, mode='same')
+        for ind in range(A.shape[0]):
+            mask[ind, :] = np.convolve(mask[ind, :], tap, mode='same')
         nmask = np.ones_like(D)
         nmask[mask < 1.] = 0
     else:
@@ -161,8 +239,8 @@ def mute(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
     # apply taper
     tap = 2. - (np.cos(np.arange(ntap, dtype=float)/ntap*2.*np.pi) + 1.)
     tap /= ntap
-    for ind in range(A.shape[1]):
-        nmask[:, ind] = np.convolve(nmask[:, ind], tap, mode='same')
+    for ind in range(A.shape[0]):
+        nmask[ind, :] = np.convolve(nmask[ind, :], tap, mode='same')
 
     # mute data with tapered mask
     A *= nmask
@@ -187,10 +265,10 @@ def normalizeStandardDeviation(
     :rtype: numpy.ndarray
     :return: normalized time series data
     """
-    std = np.std(A, axis=0)
+    std = np.std(A, axis=-1)
     # avoid creating nans or Zerodivisionerror
     std[np.where(std == 0)] = 1
-    A /= np.tile(std, (A.shape[0], 1))
+    A = (A.T/std).T
     return A
 
 
@@ -249,8 +327,8 @@ def taper(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
         func = getattr(signal, args['type'])
     args = deepcopy(args)
     args.pop('type')
-    tap = func(A.shape[0], **args)
-    A *= np.tile(np.atleast_2d(tap).T, (1, A.shape[1]))
+    tap = func(A.shape[1], **args)
+    A *= np.tile(np.atleast_2d(tap), (A.shape[0], 1))
     return A
 
 
@@ -303,11 +381,17 @@ def TDnormalization(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
     window = (
         np.ones(int(np.ceil(args['windowLength'] * params['sampling_rate'])))
         / np.ceil(args['windowLength']*params['sampling_rate']))
-    for ind in range(B.shape[1]):
-        B[:, ind] = np.convolve(B[:, ind], window, mode='same')
-        B[:, ind] = np.convolve(B[::-1, ind], window, mode='same')[::-1]
+    if len(B.shape) == 1:
+        B = np.convolve(B, window, mode='same')
+        B = np.convolve(B[::-1], window, mode='same')[::-1]
         # damping factor
-        B[:, ind] += np.max(B[:, ind])*1e-6
+        B += np.max(B)*1e-6
+    else:
+        for ind in range(B.shape[0]):
+            B[ind, :] = np.convolve(B[ind], window, mode='same')
+            B[ind, :] = np.convolve(B[ind, ::-1], window, mode='same')[::-1]
+            # damping factor
+            B[ind, :] += np.max(B[ind, :])*1e-6
     # normalization
     A /= np.sqrt(B)
     return A
@@ -343,12 +427,11 @@ def TDfilter(A: np.ndarray, args: dict, params: dict) -> np.ndarray:
     func = func_from_str('obspy.signal.filter.%s' % args['type'])
     args = deepcopy(args)
     args.pop('type')
-    # filtering in obspy.signal is done along the last dimension that why .T
-    A = func(A.T, df=params['sampling_rate'], **args).T
+    A = func(A, df=params['sampling_rate'], **args)
     return A
 
 
-def zeroPadding(A: np.ndarray, args: dict, params: dict, axis=0) -> np.ndarray:
+def zeroPadding(A: np.ndarray, args: dict, params: dict, axis=1) -> np.ndarray:
     """
     Append zeros to the traces
 
@@ -382,14 +465,13 @@ def zeroPadding(A: np.ndarray, args: dict, params: dict, axis=0) -> np.ndarray:
     """
     if A.ndim > 2 or axis > 1:
         raise NotImplementedError('Only two-dimensional arrays are supported.')
+    if A.ndim == 1 and len(A) == 0:
+        raise ValueError('Input Array is empty')
     npts = A.shape[axis]
     if A.ndim == 2:
         ntrc = A.shape[axis-1]
     elif A.ndim == 1:
         ntrc = 1
-    if not ntrc or not npts:
-        raise ValueError('Input Array is empty')
-
     if args['type'] == 'nextFastLen':
         N = next_fast_len(npts)
     elif args['type'] == 'avoidWrapAround':
