@@ -10,7 +10,7 @@ Manages the file format and class for correlations.
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Friday, 16th April 2021 03:21:30 pm
-Last Modified: Monday, 3rd April 2023 02:36:49 pm
+Last Modified: Tuesday, 4th April 2023 05:04:21 pm
 '''
 import ast
 import fnmatch
@@ -248,8 +248,7 @@ omitted." % path, category=UserWarning)
         path = '/'.join(path.split('/')[:-3])
         for ch in channel:
             for match in fnmatch.filter(self[path].keys(), ch):
-                out[match] = [
-                    k[:-7] for k in list(self['/'.join([path, match])].keys())]
+                out[match] = list(self['/'.join([path, match])].keys())
         return out
 
     def get_available_channels(
@@ -450,7 +449,7 @@ def co_to_hdf5(co: dict) -> dict:
     remk = [
         'subdir', 'read_start', 'read_end', 'read_len', 'read_inc',
         'combination_method', 'combinations', 'starttime',
-        'xcombinations']
+        'xcombinations', 'preprocess_subdiv']
     for key in remk:
         coc.pop(key, None)
         coc['corr_args'].pop('combinations', None)
