@@ -8,23 +8,26 @@
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Thursday, 18th February 2021 02:30:02 pm
-Last Modified: Tuesday, 6th December 2022 04:35:41 pm
+Last Modified: Friday, 25th August 2023 02:37:05 pm
 '''
 
 import fnmatch
 import os
 import datetime
 import glob
-from typing import List, Tuple
+from typing import List, Tuple, Iterator
 import warnings
 import re
 
+import numpy as np
 from obspy.clients.fdsn import Client as rClient
 from obspy.clients.fdsn.header import FDSNNoDataException
 from obspy.clients.filesystem.sds import Client as lClient
 from obspy import read_inventory, UTCDateTime, read, Stream, Inventory
 from obspy.clients.fdsn.mass_downloader import RectangularDomain, \
     Restrictions, MassDownloader
+
+from seismic.utils.raw_analysis import spct_series_welch
 
 
 SDS_FMTSTR = os.path.join(
@@ -115,6 +118,8 @@ class Store_Client(object):
         mdl.download(
             domain, restrictions, mseed_storage=self._get_mseed_storage,
             stationxml_storage=self.inv_dir)
+        # update inventory
+        self.read_inventory()
 
     def get_waveforms(
         self, network: str, station: str, location: str, channel: str,
@@ -164,7 +169,7 @@ class Store_Client(object):
         return statlist
 
     def _get_mseed_storage(
-        self, network: 'str', station: 'str', location: str,
+        self, network: str, station: str, location: str,
         channel: str, starttime: UTCDateTime,
             endtime: UTCDateTime) -> bool or str:
 
@@ -222,6 +227,28 @@ class Store_Client(object):
             warnings.warn(e, UserWarning)
             return (None, None)
         return (starttime, endtime)
+
+    def _translate_wildcards(self, network: str, station: str) -> List[
+            List[str]]:
+        """
+        Look up network and station names in database, so that wildcards
+        can be estimated into number of stations.
+
+        :param network: network string, may contain wildcards
+        :type network: str
+        :param station: station string, may contain wildcards
+        :type station: str
+        :return: _description_
+        :rtype: _type_
+        """
+        dirlist = glob.glob(
+            os.path.join(self.sds_root, '*', network, station, '*'))
+        nets = [os.path.basename(
+            os.path.dirname(os.path.dirname(i))) for i in dirlist]
+        stats = [os.path.basename(os.path.dirname(i)) for i in dirlist]
+        chans = [os.path.basename(i).split('.')[0] for i in dirlist]
+        # Create one nested list
+        return [[nets[i], stats[i], chans[i]] for i in range(len(dirlist))]
 
     def _load_remote(
         self, network: str, station: str, location: str, channel: str,
@@ -348,13 +375,93 @@ class Store_Client(object):
 
     def _write_inventory(self, ninv: Inventory):
         """write the inventory information"""
+        # Save inventory by station name, like this no old information
+        # can be deleted
         if self.inventory is not None:
             self.inventory += ninv
         else:
             self.inventory = ninv
-        self.inventory.write(
-            os.path.join(self.inv_dir, 'inventory.xml'),
-            format="STATIONXML", validate=True)
+        fname = os.path.join(
+            self.inv_dir, f'{ninv[0].code}.{ninv[0][0].code}.xml')
+        self.inventory.write(fname, format="STATIONXML", validate=True)
+
+    def _generate_time_windows(
+        self, network: str, station: str, channel: str, starttime: UTCDateTime,
+            endtime: UTCDateTime, increment: int = 86400) -> Iterator[Stream]:
+        """
+        Generates time windows with the requested increment from the requested
+        station.
+
+        :param network: network code
+        :type network: str
+        :param station: station code
+        :type station: str
+        :param channel: Channel Code
+        :type channel: str
+        :param starttime: Starttime of the first window
+        :type starttime: UTCDateTime
+        :param endtime: Endtime of the last window
+        :type endtime: UTCDateTime
+        :param increment: increment and window length between the windows,
+            defaults to 86400
+        :type increment: int, optional
+        :yield: Stream with every window
+        :rtype: Iterator[Stream]
+        """
+        starttimes = starttime.timestamp + np.arange(
+            0, endtime.timestamp-starttime.timestamp, increment)
+        for start in starttimes:
+            end = UTCDateTime(start + increment)
+            st = self._load_local(
+                network, station, '*', channel, UTCDateTime(start), end,
+                attach_response=True, _check_times=False)
+            yield st
+
+    def compute_spectrogram(
+        self, network: str, station: str, channel: str, starttime: UTCDateTime,
+        endtime: UTCDateTime, win_len: int, freq_max: float = 25,
+            read_increment: int = 86400) -> Tuple[
+                np.ndarray, np.ndarray, np.ndarray]:
+        """
+        Computes a time series of spectrograms for the requested station and
+        channel.
+
+        # Enter plotting function here later
+        .. seealso:: Use function :func:`~seismic.plot` to generate
+
+        .. note::
+            MPI support since version 0.4.2. Then the result is only
+            available on rank 0. All other ranks return None. (to save RAM)
+
+        :param network: network code
+        :type network: str
+        :param station: station code
+        :type station: str
+        :param channel: channel code
+        :type channel: str
+        :param starttime: starttime of the first window
+        :type starttime: UTCDateTime
+        :param endtime: endtime of the last window
+        :type endtime: UTCDateTime
+        :param win_len: Length of each time window in seconds
+        :type win_len: int
+        :param freq_max: Maximum frequency to compute, defaults to 25.
+            Also determines the Nyquist frequency.
+        :type freq_max: float, optional
+        :param read_increment: Increment to read data in - does not
+            influence the final result but has to be >= win_len,
+            defaults to 86400
+        :type read_increment: int, optional
+        :return: Frequency vector, time vector and spectrogram matrix
+        :rtype: Tuple[np.ndarray, np.ndarray, np.ndarray]
+        """
+        if read_increment < win_len:
+            raise ValueError(
+                'read_increment must be >= win_len, got {} and {}'.format(
+                    read_increment, win_len))
+        data_gen = self._generate_time_windows(
+            network, station, channel, starttime, endtime, read_increment)
+        return spct_series_welch(data_gen, win_len, freq_max)
 
 
 class FS_Client(object):
