@@ -7,7 +7,7 @@
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Thursday, 27th May 2021 04:27:14 pm
-Last Modified: Monday, 3rd April 2023 02:43:12 pm
+Last Modified: Friday, 25th August 2023 02:29:01 pm
 '''
 from copy import deepcopy
 import unittest
@@ -22,7 +22,7 @@ from obspy.core.inventory.inventory import read_inventory
 import yaml
 
 from seismic.correlate import correlate
-from seismic.correlate.stream import CorrStream
+from seismic.correlate.stream import CorrStream, CorrTrace
 from seismic.trace_data.waveform import Store_Client
 
 
@@ -37,6 +37,7 @@ class TestCorrrelator(unittest.TestCase):
             'params_example.yaml')
         with open(self.param_example) as file:
             self.options = yaml.load(file, Loader=yaml.FullLoader)
+        self.options['co']['preprocess_subdiv'] = True
 
     @mock.patch('seismic.correlate.correlate.yaml.load')
     @mock.patch('builtins.open')
@@ -47,6 +48,7 @@ class TestCorrrelator(unittest.TestCase):
         yaml_mock.return_value = self.options
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = []
+        sc_mock._translate_wildcards.return_value = []
         c = correlate.Correlator(sc_mock, self.param_example)
         self.assertDictEqual(self.options['co'], c.options)
         mkdir_calls = [
@@ -74,8 +76,10 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['network'] = ['bla']
         options['net']['station'] = ['blub']
         sc_mock = mock.Mock(Store_Client)
-        sc_mock.get_available_stations.return_value = []
+        sc_mock._translate_wildcards.return_value = [['bla', 'blub', 'E']]
         c = correlate.Correlator(sc_mock, options)
+        sc_mock._translate_wildcards.assert_called_once_with('bla', 'blub')
+        sc_mock.get_available_stations.assert_not_called()
         self.assertDictEqual(self.options['co'], c.options)
         self.assertListEqual(c.station, [['bla', 'blub']])
 
@@ -106,7 +110,7 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['network'] = '*'
         options['net']['station'] = '*'
         sc_mock = mock.Mock(Store_Client)
-        sc_mock.get_available_stations.return_value = [['lala', 'lolo']]
+        sc_mock._translate_wildcards.return_value = [['lala', 'lolo', 'E']]
         c = correlate.Correlator(sc_mock, options)
         self.assertListEqual(c.station, [['lala', 'lolo']])
 
@@ -119,10 +123,11 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['network'] = 'lala'
         options['net']['station'] = '*'
         sc_mock = mock.Mock(Store_Client)
-        sc_mock.get_available_stations.return_value = [['lala', 'lolo']]
+        sc_mock._translate_wildcards.side_effect = (
+            [['lala', 'lolo', 'E']], [['lala', 'lolo', 'Z']])
         c = correlate.Correlator(sc_mock, options)
         self.assertListEqual(c.station, [['lala', 'lolo']])
-        sc_mock.get_available_stations.assert_called_once_with('lala')
+        sc_mock._translate_wildcards.assert_called_once_with('lala', '*')
 
     @mock.patch('builtins.open')
     @mock.patch('seismic.correlate.correlate.logging')
@@ -134,9 +139,12 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['station'] = '*'
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [['lala', 'lolo']]
+        sc_mock._translate_wildcards.side_effect = (
+            [['lala', 'lala', 'E']],
+            [['lolo', 'lolo', 'E']])
         c = correlate.Correlator(sc_mock, options)
         self.assertListEqual(
-            c.station, [['lala', 'lolo'], ['lala', 'lolo']])
+            c.station, [['lala', 'lala'], ['lolo', 'lolo']])
         calls = [mock.call('lala'), mock.call('lolo')]
         sc_mock.get_available_stations.assert_has_calls(calls)
 
@@ -162,7 +170,12 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['network'] = ['lala', 'lolo']
         options['net']['station'] = ['le', 'li']
         sc_mock = mock.Mock(Store_Client)
+        sc_mock._translate_wildcards.side_effect = (
+            [['lala', 'le', 'E']], [['lolo', 'li', 'Z']])
         c = correlate.Correlator(sc_mock, options)
+        sc_mock._translate_wildcards.assert_has_calls([
+            mock.call('lala', 'le'), mock.call('lolo', 'li')])
+        sc_mock.get_available_stations.assert_not_called()
         self.assertListEqual(
             c.station, [['lala', 'le'], ['lolo', 'li']])
 
@@ -190,8 +203,8 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['station'] = '*'
         options['co']['combination_method'] = 'betweenStations'
         sc_mock = mock.Mock(Store_Client)
-        sc_mock.get_available_stations.return_value = [
-            ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
 
         sc_mock.select_inventory_or_load_remote.side_effect = [
@@ -200,12 +213,12 @@ class TestCorrrelator(unittest.TestCase):
         c.find_interstat_dist(10000)
         sc_mock.read_inventory.assert_called_once()
         sc_mock.select_inventory_or_load_remote.assert_called_with(
-            'lala', 'lili')
-        self.assertListEqual(
-            c.rcombis,
-            [
-                'lala-lala.lolo-lolo',
-                'lala-lala.lolo-lili', 'lala-lala.lili-lili'])
+            'lala', 'lolo')
+        self.assertSetEqual(
+            set(c.rcombis),
+            {
+                'lala-lala.lili-lolo',
+                'lala-lala.lili-lili', 'lala-lala.lolo-lolo'})
 
     @mock.patch('builtins.open')
     @mock.patch('seismic.correlate.correlate.logging')
@@ -217,8 +230,8 @@ class TestCorrrelator(unittest.TestCase):
         options['net']['station'] = '*'
         options['co']['combination_method'] = 'betweenComponents'
         sc_mock = mock.Mock(Store_Client)
-        sc_mock.get_available_stations.return_value = [
-            ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         with self.assertRaises(ValueError):
             c.find_interstat_dist(100)
@@ -237,6 +250,8 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         netcombs = ['AA-BB', 'AA-BB', 'AA-AA', 'AA-CC']
         statcombs = ['00-00', '00-11', '22-33', '22-44']
@@ -264,6 +279,8 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         sc_mock.read_inventory.return_value = self.inv
         cst_mock().stack.return_value = 'bla'
@@ -275,12 +292,8 @@ class TestCorrrelator(unittest.TestCase):
             c.pxcorr()
             c._generate_data.assert_called_once()
             c._pxcorr_inner.assert_called_once_with(self.st, self.inv)
-            cst_mock().stack.assert_called_with(regard_location=False)
             cst_mock().count.assert_called()
-            write_calls = [
-                mock.call('bla', 'stack_86398'),
-                mock.call('bla', 'stack_86398')
-            ]
+            write_calls = [mock.call(cst_mock()), mock.call(cst_mock())]
             c._write.assert_has_calls(write_calls)
         cst_mock().clear.assert_called()
         cst_mock().extend.assert_called_once()
@@ -296,6 +309,8 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         sc_mock.read_inventory.return_value = self.inv
         cst_mock().count.return_value = True
@@ -309,8 +324,8 @@ class TestCorrrelator(unittest.TestCase):
             cst_mock().stack.assert_not_called()
             cst_mock().count.assert_called()
             write_calls = [
-                mock.call(mock.ANY, tag='subdivision'),
-                mock.call(mock.ANY, tag='subdivision')
+                mock.call(mock.ANY),
+                mock.call(mock.ANY)
             ]
             c._write.assert_has_calls(write_calls)
         cst_mock().extend.assert_called_once()
@@ -326,10 +341,12 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         st_a_mock.return_value = (np.zeros((3, 5)), self.st)
         with mock.patch.object(c, '_pxcorr_matrix') as pxcm:
-            pxcm.return_value = (np.ones((5, 3)), np.arange(3))
+            pxcm.return_value = (np.ones((3, 5)), np.arange(5))
             cst = c._pxcorr_inner(self.st, self.inv)
             pxcm.assert_called_once()
         self.assertListEqual(
@@ -342,13 +359,17 @@ class TestCorrrelator(unittest.TestCase):
     @mock.patch('builtins.open')
     @mock.patch('seismic.correlate.correlate.logging')
     @mock.patch('seismic.correlate.correlate.os.makedirs')
-    def test_write(
+    def test_write_one_file(
             self, makedirs_mock, logging_mock, open_mock, dbh_mock):
         options = deepcopy(self.options)
         options['co']['combinations'] = [(0, 0), (0, 1), (0, 2)]
+        options['save_comps_separately'] = False
+        options['co']['subdivision']['recombine_subdivision'] = False
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         c = correlate.Correlator(sc_mock, options)
         st2 = self.st.copy()
         cst0 = CorrStream()
@@ -359,14 +380,64 @@ class TestCorrrelator(unittest.TestCase):
         cst1 = CorrStream()
         for tr in self.st:
             cst1.append(tr)
-        c._write(st2, 'mytag')
+        c._write(st2)
         add_cst_calls = [
-            mock.call(mock.ANY, 'mytag'), mock.call(mock.ANY, 'mytag')]
+            mock.call(mock.ANY, 'subdivision'),
+            mock.call(mock.ANY, 'subdivision')]
         dbh_mock().add_correlation.assert_has_calls(add_cst_calls)
         # A bit cumbersome way to check whether all files were written
         for call in dbh_mock().add_correlation.call_args_list:
             self.assertEqual(3, call[0][0].count())
 
+    @mock.patch('seismic.db.corr_hdf5.DBHandler')
+    @mock.patch('builtins.open')
+    @mock.patch('seismic.correlate.correlate.logging')
+    @mock.patch('seismic.correlate.correlate.os.makedirs')
+    def test_write_three_file(
+            self, makedirs_mock, logging_mock, open_mock, dbh_mock):
+        options = deepcopy(self.options)
+        options['co']['combinations'] = [(0, 0), (0, 1), (0, 2)]
+        options['save_comps_separately'] = True
+        options['co']['subdivision']['recombine_subdivision'] = True
+        options['co']['subdivision']['delete_subdivision'] = True
+        sc_mock = mock.Mock(Store_Client)
+        sc_mock.get_available_stations.return_value = [
+            ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
+        c = correlate.Correlator(sc_mock, options)
+        cst_mock = mock.Mock(CorrStream)
+        cst_mock2 = mock.Mock(CorrStream)
+        cst_mock.select.return_value = cst_mock2
+        # make cst_mock iterable
+        cst_mock.__iter__ = mock.Mock(return_value=iter(self.st))
+        c._write(cst_mock)
+        select_calls = [
+            mock.call(network='BW', station='RJOB', channel='EHE'),
+            mock.call().stack(regard_location=False),
+            mock.call().clear(),
+            mock.call().count(),
+            mock.call(network='BW', station='RJOB', channel='EHN'),
+            mock.call().stack(regard_location=False),
+            mock.call().clear(),
+            mock.call().count(),
+            mock.call(network='BW', station='RJOB', channel='EHZ'),
+            mock.call().stack(regard_location=False),
+            mock.call().clear(),
+            mock.call().count()]
+        cst_mock.select.assert_has_calls(select_calls)
+        # Note that subdivision is still called because clear doesn't
+        # have a defined effect on a mock
+        add_cst_calls = [
+            mock.call(mock.ANY, 'subdivision'),
+            mock.call(mock.ANY, 'stack_86398'),
+            mock.call(mock.ANY, 'subdivision'),
+            mock.call(mock.ANY, 'stack_86398'),
+            mock.call(mock.ANY, 'subdivision'),
+            mock.call(mock.ANY, 'stack_86398')]
+        dbh_mock().add_correlation.assert_has_calls(add_cst_calls)
+
+    @mock.patch('seismic.utils.miic_utils.resample_or_decimate')
     @mock.patch('seismic.correlate.correlate.calc_cross_combis')
     @mock.patch('seismic.correlate.correlate.preprocess_stream')
     @mock.patch('seismic.correlate.correlate.mu.get_valid_traces')
@@ -375,7 +446,7 @@ class TestCorrrelator(unittest.TestCase):
     @mock.patch('seismic.correlate.correlate.os.makedirs')
     def test_generate(
         self, makedirs_mock, logging_mock, open_mock, gvt_mock,
-            ppst_mock, ccc_mock):
+            ppst_mock, ccc_mock, rod_mock):
         options = deepcopy(self.options)
         options['co']['subdivision']['corr_inc'] = 5
         options['co']['subdivision']['corr_len'] = 5
@@ -383,9 +454,12 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         sc_mock.inventory = self.inv
         sc_mock._load_local.return_value = self.st
         ppst_mock.return_value = self.st
+        rod_mock.return_value = self.st
         c = correlate.Correlator(sc_mock, options)
         c.station = [['AA', '00'], ['AA', '22'], ['AA', '33'], ['BB', '00']]
         ostart = None
@@ -395,6 +469,7 @@ class TestCorrrelator(unittest.TestCase):
                 # the right place
                 self.assertAlmostEqual(win[0].stats.starttime - ostart, 5)
             ostart = win[0].stats.starttime
+            # The last one could be shorter
             self.assertAlmostEqual(
                 win[0].stats.endtime-win[0].stats.starttime, 5, 1)
 
@@ -410,6 +485,8 @@ class TestCorrrelator(unittest.TestCase):
         sc_mock = mock.Mock(Store_Client)
         sc_mock.get_available_stations.return_value = [
             ['lala', 'lolo'], ['lala', 'lili']]
+        sc_mock._translate_wildcards.return_value = [
+            ['lala', 'lolo', 'E'], ['lala', 'lili', 'Z']]
         sc_mock.select_inventory_or_load_remote.return_value = self.inv
         sc_mock._load_local.return_value = self.st
         c = correlate.Correlator(sc_mock, options)
@@ -422,10 +499,11 @@ class TestCorrrelator(unittest.TestCase):
             {'starttime': [tr.stats.starttime for tr in self.st],
                 'sampling_rate': self.st[0].stats.sampling_rate})
         # 13 is the fft size
-        shape = (101, 25)  # npts,ntrcs
-        ftshape = (51, 25)
+        shape = (25, 101)
+        ftshape = (25, 51)
         return_func = mock.MagicMock(side_effect=[
-            np.ones(shape), np.ones(ftshape)*3])
+            np.ones(shape, dtype=np.float32),
+            np.ones(ftshape, dtype=np.float32)])
         ffs_mock.return_value = return_func
         zp_mock.return_value = np.ones(shape)*2
         C, startlags = c._pxcorr_matrix(np.zeros(shape))
@@ -437,13 +515,13 @@ class TestCorrrelator(unittest.TestCase):
         self.assertListEqual([], return_func.call_args_list[0][0][1])
         # Check if the fft worked
         np.testing.assert_array_almost_equal(
-            np.fft.rfft(np.ones(shape)*2, axis=0),
+            np.fft.rfft(np.ones(shape)*2),
             return_func.call_args_list[1][0][0])
         np.testing.assert_array_equal(-np.ones((3,)), startlags)
         # Correlation should be one in the middle
         # Length is 51,3 as above
-        expC = np.zeros((51, 3))
-        expC[25] += 1
+        expC = np.zeros((3, 51))
+        expC[:, 25] += 1
         np.testing.assert_array_almost_equal(C, expC, decimal=2)
 
 
@@ -453,7 +531,7 @@ class TestStToNpArray(unittest.TestCase):
 
     def test_result_shape(self):
         A, _ = correlate.st_to_np_array(self.st, self.st[0].stats.npts)
-        self.assertEqual(A.shape, (self.st[0].stats.npts, self.st.count()))
+        self.assertEqual(A.shape, (self.st.count(), self.st[0].stats.npts))
 
     def test_deleted_data(self):
         _, st = correlate.st_to_np_array(self.st, self.st[0].stats.npts)
@@ -464,7 +542,7 @@ class TestStToNpArray(unittest.TestCase):
     def test_result(self):
         A, _ = correlate.st_to_np_array(self.st.copy(), self.st[0].stats.npts)
         for ii, tr in enumerate(self.st):
-            self.assertTrue(np.allclose(tr.data, A[:, ii]))
+            self.assertTrue(np.allclose(tr.data, A[ii]))
 
 
 class TestCompareExistingData(unittest.TestCase):
@@ -486,7 +564,7 @@ class TestCompareExistingData(unittest.TestCase):
                 '%s.%s' % (tr1.stats.network, tr1.stats.station)][
             '%s-%s' % (
                 tr0.stats.channel, tr1.stats.channel)] = [
-                    tr0.stats.starttime.format_fissures()[:-7]]
+                    tr0.stats.starttime.format_fissures()]
         self.ex_d = ex_d
         self.tr0 = tr0
         self.tr1 = tr1
@@ -622,7 +700,7 @@ class TestSortCombnameAlphabetically(unittest.TestCase):
         net1 = 'B'
         stat0 = 'Z'
         stat1 = 'C'
-        exp_result = ([net0, net1], [stat0, stat1])
+        exp_result = ([net0, net1], [stat0, stat1], ['', ''])
         self.assertEqual(
             correlate.sort_comb_name_alphabetically(net0, stat0, net1, stat1),
             exp_result)
@@ -632,9 +710,35 @@ class TestSortCombnameAlphabetically(unittest.TestCase):
         net1 = 'A'
         stat0 = 'Z'
         stat1 = 'C'
-        exp_result = ([net1, net0], [stat1, stat0])
+        exp_result = ([net1, net0], [stat1, stat0], ['', ''])
         self.assertEqual(
             correlate.sort_comb_name_alphabetically(net0, stat0, net1, stat1),
+            exp_result)
+
+    def test_between_comps(self):
+        net0 = 'A'
+        net1 = 'A'
+        stat0 = 'Z'
+        stat1 = 'Z'
+        cha0 = 'B'
+        cha1 = 'A'
+        exp_result = ([net0, net1], [stat0, stat1], [cha1, cha0])
+        self.assertEqual(
+            correlate.sort_comb_name_alphabetically(
+                net0, stat0, net1, stat1, cha0, cha1),
+            exp_result)
+
+    def test_between_comps_no_flip(self):
+        net0 = 'A'
+        net1 = 'A'
+        stat0 = 'Z'
+        stat1 = 'Z'
+        cha1 = 'B'
+        cha0 = 'A'
+        exp_result = ([net0, net1], [stat0, stat1], [cha0, cha1])
+        self.assertEqual(
+            correlate.sort_comb_name_alphabetically(
+                net0, stat0, net1, stat1, cha0, cha1),
             exp_result)
 
     def test_wrong_arg_type(self):
@@ -729,24 +833,6 @@ class TestPreProcessStream(unittest.TestCase):
         self.assertEqual(
             correlate.preprocess_stream(Stream(), **self.kwargs), Stream())
 
-    def test_wrong_sr(self):
-        x = np.random.randint(1, 100)
-        sr = self.st[0].stats.sampling_rate + x
-        kwargs = deepcopy(self.kwargs)
-        kwargs['sampling_rate'] = sr
-        with self.assertRaises(ValueError):
-            correlate.preprocess_stream(self.st, **kwargs)
-
-    def test_decimate(self):
-        st = correlate.preprocess_stream(self.st.copy(), **self.kwargs)
-        self.assertEqual(25, st[0].stats.sampling_rate)
-
-    def test_resample(self):
-        kwargs = deepcopy(self.kwargs)
-        kwargs['sampling_rate'] = 23
-        st = correlate.preprocess_stream(self.st.copy(), **kwargs)
-        self.assertEqual(23, st[0].stats.sampling_rate)
-
     def test_pad(self):
         kwargs = deepcopy(self.kwargs)
         kwargs['startt'] -= 10
@@ -770,7 +856,7 @@ class TestPreProcessStream(unittest.TestCase):
     def test_discard_short(self):
         kwargs = deepcopy(self.kwargs)
         kwargs['startt'] += 15
-        kwargs['endt'] -= 15
+        kwargs['endt'] -= 5
         kwargs['subdivision']['corr_len'] = 200
         st = correlate.preprocess_stream(self.st.copy(), **kwargs)
         self.assertFalse(st.count())
