@@ -8,7 +8,7 @@
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Tuesday, 20th July 2021 04:00:46 pm
-Last Modified: Monday, 16th January 2023 11:14:10 am
+Last Modified: Wednesday, 28th June 2023 02:02:47 pm
 '''
 import unittest
 from unittest.mock import patch
@@ -28,7 +28,7 @@ class TestFDFilter(unittest.TestCase):
         freqs = np.fft.fftfreq(100, d=1/100)
         self.params['freqs'] = freqs
         A = np.fft.fft(np.cos(np.linspace(0, 2*np.pi, 100)))
-        A = np.tile(A, (2, 1)).T
+        A = np.tile(A, (2, 1))
         out = ppfd.FDfilter(A, args, self.params)
         np.testing.assert_allclose(out.real, 0, atol=1e-4)
 
@@ -37,12 +37,14 @@ class TestFDSignBitNormalisation(unittest.TestCase):
     # Not much to test here
     def test_result(self):
         np.random.seed(2)
-        dim = (np.random.randint(200, 766), np.random.randint(2, 44))
+        dim = (np.random.randint(2, 6), np.random.randint(200, 766))
         A = np.random.random(dim)-.5
-        Aft = np.fft.rfft(A, axis=0)
-        expected_result = np.fft.rfft(np.sign(A), axis=0)
-        self.assertTrue(np.allclose(
-            expected_result, ppfd.FDsignBitNormalization(Aft, {}, {})))
+        B = np.fft.rfft(A)
+        # to acccount for inaccuracies in the FFT and the fact
+        # that the imaginary part will be ommited
+        expected_result = np.fft.rfft(np.sign(np.fft.irfft(B)))
+        np.testing.assert_allclose(
+            expected_result, ppfd.FDsignBitNormalization(B, {}, {}))
 
 
 class TestSpectralWhitening(unittest.TestCase):
@@ -54,7 +56,7 @@ class TestSpectralWhitening(unittest.TestCase):
         # Again so straightforward that I wonder whether it makes sense
         # to test this
         expected = self.A/abs(self.A)
-        expected[0, :] = 0.j
+        expected[:, 0] = 0.j
         self.assertTrue(np.allclose(
             expected, ppfd.spectralWhitening(self.A, {}, {})))
 
@@ -64,9 +66,10 @@ class TestSpectralWhitening(unittest.TestCase):
                 np.ones((5, 5)), {'joint_norm': True}, {})
 
     def test_joint_norm(self):
-        A = np.tile(np.random.random(55,) + np.random.random(55,)*1j, (3, 1)).T
+        A = np.tile(np.random.random(55,) + np.random.random(55,)*1j, (3, 1))
+        A -= .5
         expected = A/abs(A)
-        expected[0, :] = 0.j
+        expected[:, 0] = 0.j
         out = ppfd.spectralWhitening(
             A, {'joint_norm': True}, {})
         self.assertTrue(np.allclose(
