@@ -8,7 +8,7 @@
    Peter Makus (makus@gfz-potsdam.de)
 
 Created: Monday, 16th January 2023 11:07:27 am
-Last Modified: Wednesday, 22nd February 2023 03:47:10 pm
+Last Modified: Friday, 28th July 2023 03:33:13 pm
 '''
 
 import unittest
@@ -28,50 +28,61 @@ class TestProbability(unittest.TestCase):
         t = np.random.rand() + 5
         c = dist/t - np.random.rand()
         # mf path should not matter
-        self.assertEqual(spt.probability(dist, t, c, 10), 0)
+        self.assertEqual(spt.probability(dist, t, c, 10, 1e-6), 0)
 
     def test_vel0(self):
         with self.assertRaises(ZeroDivisionError):
-            spt.probability(1, 1, 0, 10), 1
+            spt.probability(1, 1, 0, 10, 1e-9), 1
 
-    def test_prob1(self):
+    def test_ballistic(self):
         # distance=t*v
         dist = 3
         t = 5
         c = dist/t
         # mf path should not matter
-        self.assertEqual(spt.probability(dist, t, c, 10), 1)
+        self.assertGreater(spt.probability(dist, t, c, 10, 1e-6), 0)
 
-    def test_prob1_array(self):
+    def test_ballistic_array(self):
         # distance=t*v
         dist = np.ones(40)*2
         t = 5
-        c = dist/t
+        c = 2/t
         # mf path should not matter
-        np.testing.assert_almost_equal(spt.probability(dist, t, c, 10), 1)
+        np.testing.assert_array_less(
+            0, spt.probability(dist, t, c, 10, 1e-6))
+
+    def test_ballistic_array2(self):
+        dist = np.ones(40)*2
+        t = np.arange(6)
+        c = 2/5
+        # mf path should not matter
+        prob = spt.probability(dist, t, c, 10, 1e-6)
+        np.testing.assert_array_less(0, prob[-1, :, :])
+        # before the ballistic arrival, the probability should be 0
+        np.testing.assert_array_almost_equal(prob[0:-1], 0)
 
     def test_prob_rand(self):
         # in any other case this should be 0 < probability < 1
         dist = np.random.rand() + 3
         t = np.random.rand() + 5
-        c = dist/t + .01 + np.random.rand()
+        c = 2/t + .01 + np.random.rand()
         # mf path should not matter
-        self.assertGreater(spt.probability(dist, t, c, 10), 0)
-        self.assertLess(spt.probability(dist, t, c, 10), 1)
+        self.assertGreater(spt.probability(dist, t, c, 10, c*t/2), 0)
+        self.assertLess(spt.probability(dist, t, c, 10, c*t/2), 1)
 
     def test_dist_neg(self):
         dist = -3
         t = 5
         c = dist/t
         with self.assertRaises(ValueError):
-            spt.probability(dist, t, c, 10)
+            spt.probability(dist, t, c, 10, c*t/2)
 
     def test_t_neg(self):
         dist = 3
         t = -1
         c = 1
         with self.assertRaises(ValueError):
-            spt.probability(dist, t, c, 10)
+            spt.probability(dist, t, c, 10, 1e-3)
 
 
 class TestComputeGridDist(unittest.TestCase):
@@ -93,13 +104,13 @@ class TestSensitivityKernel(unittest.TestCase):
     @mock.patch('seismic.monitor.spatial.compute_grid_dist')
     @mock.patch('seismic.monitor.spatial.probability')
     def test_result(self, pbb_mock, cgd_mock):
-        dt = .1
+        dt = 1
         s1 = s2 = np.array([5, 5])
         x = y = np.arange(10)
         t = 20
         vel = mf_path = 1
         cgd_mock.return_value = np.zeros((10, 10))
-        pbb_mock.return_value = np.ones((10, 10))
+        pbb_mock.side_effect = [1] + [np.ones((10, 10))]*50
         K = spt.sensitivity_kernel(s1, s2, x, y, t, dt, vel, mf_path)
         cgd_calls = [
             mock.call(x, y, s1[0], s1[1]),
@@ -107,16 +118,37 @@ class TestSensitivityKernel(unittest.TestCase):
         ]
         cgd_mock.assert_has_calls(cgd_calls)
         # To be continued here
-        self.assertEqual(pbb_mock.call_count, 403)
+        self.assertEqual(pbb_mock.call_count, 43)
 
-        calls = [mock.call(0, 20, 1, 1, atol=dt/2)]
+        calls = [mock.call(0, 20, 1, 1, atol=dt*vel/2)]
         for tt in np.arange(0, t + dt, dt):
             calls.extend([
-                mock.call(mock.ANY, tt, vel, mf_path, atol=dt/2),
-                mock.call(mock.ANY, t-tt, vel, mf_path, atol=dt/2)])
+                mock.call(mock.ANY, tt, vel, mf_path, atol=dt*vel/2),
+                mock.call(mock.ANY, t-tt, vel, mf_path, atol=dt*vel/2)])
         pbb_mock.assert_has_calls(calls)
 
         np.testing.assert_almost_equal(K, np.ones((10, 10))*t)
+
+    @mock.patch('seismic.monitor.spatial.compute_grid_dist')
+    @mock.patch('seismic.monitor.spatial.probability')
+    def test_dstat_too_large(self, pbb_mock, cgd_mock):
+        dt = 1
+        s1 = np.array([5, 5])
+        s2 = np.array([5, 25])
+        x = y = np.arange(10)
+        t = 10
+        vel = mf_path = 1
+        cgd_mock.return_value = np.zeros((10, 10))
+        pbb_mock.side_effect = [0] + [np.ones((10, 10))]*50
+        K = spt.sensitivity_kernel(s1, s2, x, y, t, dt, vel, mf_path)
+        cgd_calls = [
+            mock.call(x, y, s1[0], s1[1])]
+        cgd_mock.assert_has_calls(cgd_calls)
+        # To be continued here
+        self.assertEqual(pbb_mock.call_count, 1)
+        pbb_mock.assert_called_once_with(20.0, 10, 1, 1, atol=dt*vel/2)
+
+        np.testing.assert_array_equal(K, 0)
 
     def test_x_not_monotoneous(self):
         dt = .1
@@ -146,6 +178,21 @@ class TestSensitivityKernel(unittest.TestCase):
         vel = mf_path = 1
         with self.assertRaises(ValueError):
             spt.sensitivity_kernel(s1, s2, x, y, t, dt, vel, mf_path)
+
+    @mock.patch('seismic.monitor.spatial.compute_grid_dist')
+    @mock.patch('seismic.monitor.spatial.probability')
+    def test_warn_on_large_dt(self, pbb_mock, cgd_mock):
+        dt = 1
+        s1 = s2 = np.array([5, 5])
+        x = y = np.arange(10)
+        t = 20
+        vel = mf_path = 1
+        cgd_mock.return_value = np.zeros((10, 10))
+        pbb_mock.side_effect = [1] + [np.ones((10, 10))]*50
+        with warnings.catch_warnings(record=True) as w:
+            spt.sensitivity_kernel(s1, s2, x, y, t, dt, vel, mf_path)
+        self.assertEqual(len(w), 1)
+        self.assertTrue(issubclass(w[0].category, UserWarning))
 
 
 class TestDataVariance(unittest.TestCase):
@@ -191,6 +238,14 @@ class TestDataVariance(unittest.TestCase):
         with self.assertRaises(ValueError):
             spt.data_variance(corr, bw, tw, freq_c)
 
+    def test_raise_error_for_tw1_greater_than_tw0(self):
+        corr = np.ones(10)
+        bw = 2
+        tw = (10, 5)
+        freq_c = 3
+        with self.assertRaises(ValueError):
+            spt.data_variance(corr, bw, tw, freq_c)
+
 
 class TestComputeCM(unittest.TestCase):
     def test_result(self):
@@ -221,6 +276,14 @@ class TestGeo2Cart(unittest.TestCase):
         self.assertTrue(np.all(out[0] < 0))
         self.assertTrue(np.all(out[1] >= 0))
 
+    def test_result_lonneg_float(self):
+        lat = 10.
+        lon = -10
+        lat0 = -5
+        out = spt.geo2cart(lat, lon, lat0)
+        self.assertTrue(np.all(out[0] < 0))
+        self.assertTrue(np.all(out[1] >= 0))
+
     def test_result_len1(self):
         # Just making sure that it still remains a np array
         # otherwise this would introduce bugs
@@ -236,11 +299,11 @@ class TestGeo2Cart(unittest.TestCase):
 
 class TestDVGrid(unittest.TestCase):
     def setUp(self):
-        self.dvg = spt.DVGrid(0, 0, 1, 9, 9)
+        dt = vel = mf_path = 1
+        self.dvg = spt.DVGrid(0, 0, 1, 9, 9, dt, vel, mf_path)
 
     def test_forward_model_from_dv(self):
         mod = np.ones((10, 10))
-        dt = vel = mf_path = 1
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -252,46 +315,41 @@ class TestDVGrid(unittest.TestCase):
                 None)
             st_mock.return_value, _ = np.meshgrid(
                 np.arange(100), np.arange(100))
-            out = self.dvg.forward_model(mod, dt, vel, mf_path, dvmock, utc=0)
+            out = self.dvg.forward_model(mod, dvmock, utc=0)
             ex_mock.assert_called_once_with(dvmock, 0)
             st_mock.assert_called_once_with(
-                'slat0', 'slon0', 'slat1', 'slon1', 5, dt, vel, mf_path)
+                'slat0', 'slon0', 'slat1', 'slon1', 5)
         np.testing.assert_allclose(out, np.ones(100)*4950)
 
     def test_forward_model_from_kwargs(self):
         mod = np.ones((10, 10))
-        dt = vel = mf_path = 1
         with mock.patch.object(
                 self.dvg, '_compute_sensitivity_kernels') as st_mock:
             st_mock.return_value, _ = np.meshgrid(
                 np.arange(100), np.arange(100))
             out = self.dvg.forward_model(
-                mod, dt, vel, mf_path, tw=(3, 7), stat0=[('slat0', 'slon0')],
+                mod, tw=(3, 7), stat0=[('slat0', 'slon0')],
                 stat1=[('slat1', 'slon1')])
             st_mock.assert_called_once_with(
-                ('slat0',), ('slon0',), ('slat1',), ('slon1',), 5, dt, vel,
-                mf_path)
+                ('slat0',), ('slon0',), ('slat1',), ('slon1',), 5)
         np.testing.assert_allclose(out, np.ones(100)*4950)
 
     def test_wrong_grid(self):
         mod = np.ones((100, 100))
-        dt = vel = mf_path = 1
         dvmock = mock.MagicMock()
         with self.assertRaises(ValueError):
-            self.dvg.forward_model(mod, dt, vel, mf_path, dvmock, utc=0)
+            self.dvg.forward_model(mod, dvmock, utc=0)
 
     def test_missing_args(self):
         mod = np.ones((10, 10))
-        dt = vel = mf_path = 1
         with self.assertRaises(TypeError):
-            self.dvg.forward_model(mod, dt, vel, mf_path)
+            self.dvg.forward_model(mod)
 
     @mock.patch('seismic.monitor.spatial.compute_cm')
     def test_compute_dv_grid_from_dv(self, cm_mock: mock.MagicMock):
         cm_mock.return_value = np.zeros((100, 100))
 
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -309,11 +367,11 @@ class TestDVGrid(unittest.TestCase):
             cd_mock.return_value = np.eye(50)
 
             out = self.dvg.compute_dv_grid(
-                dvmock, 0, dt, vel, mf_path, scaling_factor, corr_len,
+                dvmock, 0, scaling_factor, corr_len,
                 std_model, compute_resolution=False)
             ex_mock.assert_called_once_with(dvmock, 0)
             st_mock.assert_called_once_with(
-                'slat0', 'slon0', 'slat1', 'slon1', 5, dt, vel, mf_path)
+                'slat0', 'slon0', 'slat1', 'slon1', 5)
             cd_mock.assert_called_once_with(
                 mock.ANY, 3, 5, (3, 7), mock.ANY)
         cm_mock.assert_called_once_with(
@@ -324,8 +382,7 @@ class TestDVGrid(unittest.TestCase):
     def test_compute_dv_grid_from_kwargs(self, cm_mock: mock.MagicMock):
         cm_mock.return_value = np.zeros((100, 100))
 
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -344,12 +401,12 @@ class TestDVGrid(unittest.TestCase):
             cd_mock.return_value = np.eye(50)
 
             out = self.dvg.compute_dv_grid(
-                dvmock, 0, dt, vel, mf_path, scaling_factor, corr_len,
+                dvmock, 0, scaling_factor, corr_len,
                 std_model, compute_resolution=True, tw=(3, 7), freq0=3,
                 freq1=5)
             ex_mock.assert_called_once_with(dvmock, 0)
             st_mock.assert_called_once_with(
-                'slat0', 'slon0', 'slat1', 'slon1', 5, dt, vel, mf_path)
+                'slat0', 'slon0', 'slat1', 'slon1', 5)
             cd_mock.assert_called_once_with(
                 mock.ANY, 3, 5, (3, 7), mock.ANY)
             res_mock.assert_called_once()
@@ -358,8 +415,7 @@ class TestDVGrid(unittest.TestCase):
         np.testing.assert_allclose(out, np.zeros((10, 10)))
 
     def test_compute_dv_grid_dvproc_missing(self):
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -369,7 +425,7 @@ class TestDVGrid(unittest.TestCase):
                 np.zeros(50), np.ones(50), 'slat0', 'slon0', 'slat1',
                 'slon1', None, None, None)
             self.dvg.compute_dv_grid(
-                dvmock, 0, dt, vel, mf_path, scaling_factor, corr_len,
+                dvmock, 0, scaling_factor, corr_len,
                 std_model, compute_resolution=False)
             ex_mock.assert_called_once_with(dvmock, 0)
 
@@ -377,8 +433,7 @@ class TestDVGrid(unittest.TestCase):
     def test_compute_res_from_dv(self, cm_mock: mock.MagicMock):
         cm_mock.return_value = np.zeros((100, 100))
 
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -397,11 +452,10 @@ class TestDVGrid(unittest.TestCase):
             cd_mock.return_value = np.eye(50)
 
             self.dvg.compute_resolution(
-                dvmock, 0, scaling_factor, corr_len, dt, std_model, vel,
-                mf_path)
+                dvmock, 0, scaling_factor, corr_len, std_model)
             ex_mock.assert_called_once_with(dvmock, 0)
             st_mock.assert_called_once_with(
-                'slat0', 'slon0', 'slat1', 'slon1', 5, dt, vel, mf_path)
+                'slat0', 'slon0', 'slat1', 'slon1', 5)
             cd_mock.assert_called_once_with(
                 mock.ANY, 3, 5, (3, 7), mock.ANY)
         cm_mock.assert_called_once_with(
@@ -411,8 +465,7 @@ class TestDVGrid(unittest.TestCase):
     def test_compute_res_from_kwargs(self, cm_mock: mock.MagicMock):
         cm_mock.return_value = np.zeros((100, 100))
 
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -430,11 +483,11 @@ class TestDVGrid(unittest.TestCase):
                 np.arange(100), np.arange(100))
             cd_mock.return_value = np.eye(50)
             self.dvg.compute_resolution(
-                dvmock, 0, scaling_factor, corr_len, dt, std_model, vel,
-                mf_path, tw=(3, 7), freq0=3, freq1=5)
+                dvmock, 0, scaling_factor, corr_len, std_model, tw=(3, 7),
+                freq0=3, freq1=5)
             ex_mock.assert_called_once_with(dvmock, 0)
             st_mock.assert_called_once_with(
-                'slat0', 'slon0', 'slat1', 'slon1', 5, dt, vel, mf_path)
+                'slat0', 'slon0', 'slat1', 'slon1', 5)
             cd_mock.assert_called_once_with(
                 mock.ANY, 3, 5, (3, 7), mock.ANY)
             res_mock.assert_called_once()
@@ -442,8 +495,7 @@ class TestDVGrid(unittest.TestCase):
             scaling_factor, corr_len, std_model, mock.ANY)
 
     def test_compute_res_dvproc_missing(self):
-        dt, vel, mf_path, scaling_factor, corr_len, std_model \
-            = np.arange(6) + 1
+        scaling_factor, corr_len, std_model = [4, 5, 1]
         dvmock = mock.MagicMock()
         with (
             mock.patch.object(self.dvg, '_extract_info_dvs') as ex_mock,
@@ -453,8 +505,7 @@ class TestDVGrid(unittest.TestCase):
                 np.zeros(50), np.ones(50), 'slat0', 'slon0', 'slat1',
                 'slon1', None, None, None)
             self.dvg.compute_resolution(
-                dvmock, 0, scaling_factor, corr_len, dt, std_model, vel,
-                mf_path)
+                dvmock, 0, scaling_factor, corr_len, std_model)
             ex_mock.assert_called_once_with(dvmock, 0)
 
     def test_compute_res_int(self):
@@ -612,6 +663,8 @@ class TestDVGrid(unittest.TestCase):
         lats = np.arange(10)
         lons = np.arange(10) + 10
         dvg.statx = lats + 5
+        dvg.statlon = lons
+        dvg.statlat = lats
         dvg.staty = lons + 5
         g2c_mock.return_value = (lats, lons)
         dvg._add_stations(lats, lons)
@@ -635,7 +688,7 @@ class TestDVGrid(unittest.TestCase):
         out = self.dvg._compute_cd(skernels, freq0, freq1, tw, corrs)
         dv_mock.assert_called_once_with(np.array('c'), 2, tw, 2)
         np.testing.assert_array_almost_equal(
-            np.eye(100)*dv_mock.return_value, out)
+            np.eye(100)*dv_mock.return_value**2, out)
 
     def test_compute_dist(self):
         dist = self.dvg._compute_dist_matrix()
@@ -678,17 +731,33 @@ class TestDVGrid(unittest.TestCase):
         slon0 = 'slon0'
         slat1 = 'slat1'
         slon1 = 'slon1'
-        t = 5
-        dt = 0.1
-        vel = mf_path = 1
         g2c_mock.side_effect = ([[0], [1]], [[2], [3]])
         out = self.dvg._compute_sensitivity_kernels(
-            slat0, slon0, slat1, slon1, t, dt, vel, mf_path)
+            slat0, slon0, slat1, slon1, 1)
         calls = [
             mock.call(slat0, slon0, self.dvg.lat0),
             mock.call(slat1, slon1, self.dvg.lat0)]
         g2c_mock.assert_has_calls(calls)
         np.testing.assert_array_almost_equal(out, np.array([np.zeros((4))]))
+
+    @mock.patch('seismic.monitor.spatial.geo2cart')
+    @mock.patch('seismic.monitor.spatial.sensitivity_kernel')
+    def test_compute_sensitivity_kernels_already_computed(
+            self, sk_mock: mock.MagicMock, g2c_mock: mock.MagicMock):
+        slat0 = '0'
+        slon0 = '0'
+        slat1 = '1'
+        slon1 = '1'
+        self.dvg.skernels[(0, 1, 2, 3, 1)] = np.zeros((2, 2))
+        g2c_mock.side_effect = ([[0], [1]], [[2], [3]])
+        out = self.dvg._compute_sensitivity_kernels(
+            slat0, slon0, slat1, slon1, 1)
+        calls = [
+            mock.call(slat0, slon0, self.dvg.lat0),
+            mock.call(slat1, slon1, self.dvg.lat0)]
+        g2c_mock.assert_has_calls(calls)
+        sk_mock.assert_not_called()
+        np.testing.assert_array_almost_equal(out, np.array([np.zeros((2, 2))]))
 
 
 if __name__ == "__main__":
